requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-ollama
langchain-community
langgraph
chromadb
httpx
pytest
pytest-asyncio
